cleaned up the whole scheduling system
